@@ -3,17 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material/linear-progress": "^4.0.0",
-    "d3-format": "^2.0.0",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@material/linear-progress": "^14.0.0",
+    "d3-format": "^3.1.0",
     "lodash": "^4.17.15",
     "moment": "^2.26.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-redux": "^7.1.3",
-    "react-scripts": "4.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^9.1.0",
+    "react-scripts": "5.0.1",
     "react-vis": "^1.11.7",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
+    "redux": "^5.0.1",
+    "redux-thunk": "^3.1.0",
     "source-map-explorer": "^2.4.2",
     "twix": "^1.3.0"
   },
@@ -26,7 +27,7 @@
   },
   "jest": {
     "transformIgnorePatterns": [
-      "/node_modules/(?!(@material/linear-progress|react-vis)/)",
+      "/node_modules/(?!(@material|d3-.*|internmap|linear-progress|react-vis)/)",
       "\\.pnp\\.[^\\/]+$"
     ]
   },
