@@ -1,10 +1,10 @@
 import './index.css';
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import {createRoot} from 'react-dom/client';
 import {Provider} from "react-redux";
 import {createStore, applyMiddleware} from 'redux';
-import thunkMiddleware from 'redux-thunk'
+import {thunk} from 'redux-thunk';
 
 // import registerServiceWorker from './registerServiceWorker';
 import rootReducer from './store';
@@ -14,16 +14,16 @@ import App from './views/App';
 
 const store = createStore(
   rootReducer,
-  applyMiddleware(thunkMiddleware)
+  applyMiddleware(thunk)
 );
 
-
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+const onRefresh = () => store.dispatch(refreshAll());
+root.render(
   <Provider store={store}>
-    <App />
-  </Provider>,
-  document.getElementById('root'),
-  () => store.dispatch(refreshAll())
-);
+    <App onRefresh={onRefresh}/>
+  </Provider>
+)
 
 // registerServiceWorker();
